# server/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.3"
  title: titulo
  description: descrição

# during dev, should point to your local machine
host: localhost:8000
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
securityDefinitions:
 Bearer:
  description: |
   Autenticação da API se baseia nos seguintes recursos:

   - API Key
   - OAuth2

   O acesso a API ncecessita de um token JWT válido que deve ser passado no cabeçalho no campo 'Authorization'.

   Um token JWT válido é gerado pela API e retornado na resposta a rota
   /login uma vez que possua um email & senha válido.

   A seguinte sintaxe deve ser utilizada no cabeçalho das requisições 'Authorization' :
       Bearer: xxxxxx.yyyyyyy.zzzzzz
  type: apiKey
  name: Authorization
  in: header
paths:
  /login:
    x-swagger-router-controller: login
    post:
      summary: Loga o usuário
      operationId: authenticate
      tags:
        - Infraestrutura
      description: |
        Permite ao usuário logar e receber um token uma vez valida a requisição
      parameters:
        -
         in: body
         name: body
         description: Informação de email/senha
         required: true
         schema:
          $ref: '#/definitions/Login'
      responses:
        '200':
         description: Login Success
         schema:
          $ref: '#/definitions/Token'
        '400':
         description: Quando o usuário não é encontrado.
         schema:
          $ref: '#/definitions/Error'
        '403':
         description: >-
          Quando usuário existe mas possui credenciais inválidas (bad credentials) OU se o usuário não puder logar
         schema:
          $ref: '#/definitions/Error'
        '503':
         description: Erro de infraestrutura.
         schema:
          $ref: '#/definitions/Error'
...
  /api/aplicacao:
    x-swagger-router-controller: aplicacao
    get:
      description: Retorna as aplicações passíveis de serem solicitadas
      summary: Lista de aplicações cadastradas
      operationId: getAplicacoes
      tags:
        - Cadastros Auxiliares
      responses:
        "200":
          description: Sucesso
          schema:
            $ref: "#/definitions/GetAplicacoesResponse"
        default:
          description: Erro
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  Login:
    type: object
    properties:
      username:
        description: Email do usuário
        type: string
      password:
        description: Senha do usuário
        type: string
      aplicacao:
        description: Aplicação
        type: integer
  Token:
    type: object
    properties:
      token:
        type: string
  Error:
    type: object
    properties:
      message:
        type: string
        description: Retorna uma mensagem de sucesso
      error:
        type: string
        description: Retorna o código de erro. Zero se deu certo.
...
  GetAplicacoesResponse:
    required:
      - aplicacoes
    properties:
      aplicacoes:
        type: array
        items:
          type: object
          properties:
            codigo:
              type: number
            descricao:
              type: string
            mime:
              type: string
...
  ErrorResponse:
    required:
      - message
    properties:
      codret:
        type: string
      message:
        type: string